Split sensors handling so the separate file
